Escape CMAKE_CROSSCOMPILING_EMULATOR
Emscripten passes `/path/to/node --experimental-options` as CMAKE_CROSSCOMPILING_EMULATOR as a list, so `/path/to/node;--options`. This is then picked up as two list items to `set()`. Escaping with `"${CMAKE_CROSSCOMPILING_EMULATOR}"` didn't work because that resulted in the string separated by spaces, which down the line was treated as two separate options to the `CONFIGURE_COMMAND`. This patch basically escapes the whitespace that was causing trouble, and now configuration is called correctly.